tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.stroke.join.2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.stroke.join.2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.stroke.join.2</h1>
<p class="desc">Shadows are drawn for stroke joins</p>


<script>
var t = async_test("Shadows are drawn for stroke joins");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 50, 50);
  ctx.fillStyle = '#0f0';
  ctx.fillRect(50, 0, 50, 50);
  ctx.strokeStyle = '#f00';
  ctx.shadowColor = '#0f0';
  ctx.shadowOffsetX = 100;
  ctx.lineWidth = 200;
  ctx.lineJoin = 'miter';
  ctx.beginPath();
  ctx.moveTo(-200, -50);
  ctx.lineTo(-150, -50);
  ctx.lineTo(-151, -100);
  ctx.stroke();

  _assertPixel(canvas, 1,1, 0,255,0,255);
  _assertPixel(canvas, 48,48, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 98,48, 0,255,0,255);
  t.done();

});
</script>
